pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["svetlana_akaemova <akaemova.sv@mail.ru>"]
packages = [
    { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
